Use tasks.get() instead of Task.getTaskDependencies() to resolve tasks in Velocity context
Task.getTaskDependencies() will attempt to instantiate the current task,
which leads to a cyclic dependency if the property to be resolved
is used as an input to the task (need the task to resolve the inputs,
but need the inputs to construct the task). Instead, use project.tasks.get()
to get the specific tasks instance requested, assuming that the requested task
has the same release and entry values. This is a more limited version of
the old behaviour, where the first matching task among the dependencies of
the current task would be used (e.g. mux.01.TV could depend on merge.03.TV
and "$merge" would resolve to the correct task). This should not be an issue
in practice for most use cases, but this may need to be revisited at some point
in the future. Fixes #11.